Reorganize the nmagent internal package
The wireserver transport became big enough to warrant its own file
